second

+++ anto.py
@@ -4,31 +4,47 @@ import sys
 import pygame
 import requests
 
+
+def draw():
+    global map_file, ch
+    map_request = f"http://static-maps.yandex.ru/1.x/?&pt={coord}&z={z}&l=map"
+    response = requests.get(map_request)
+
+    if not response:
+        print("Ошибка выполнения запроса:")
+        print(map_request)
+        print("Http статус:", response.status_code, "(", response.reason, ")")
+        sys.exit(1)
+
+    map_file = "map.png"
+    with open(map_file, "wb") as file:
+        file.write(response.content)
+    screen = pygame.display.set_mode((600, 450))
+    # Рисуем картинку, загружаемую из только что созданного файла.
+    screen.blit(pygame.image.load(map_file), (0, 0))
+    ch = False
+
+
 coord = ','.join(input().split())
 z = int(input())
-map_request = f"http://static-maps.yandex.ru/1.x/?&pt={coord}&z={z}&l=map"
-response = requests.get(map_request)
-
-if not response:
-    print("Ошибка выполнения запроса:")
-    print(map_request)
-    print("Http статус:", response.status_code, "(", response.reason, ")")
-    sys.exit(1)
-
-map_file = "map.png"
-with open(map_file, "wb") as file:
-    file.write(response.content)
-
-# Инициализируем pygame
 pygame.init()
-screen = pygame.display.set_mode((600, 450))
-# Рисуем картинку, загружаемую из только что созданного файла.
-screen.blit(pygame.image.load(map_file), (0, 0))
-# Переключаем экран и ждем закрытия окна.
-pygame.display.flip()
-while pygame.event.wait().type != pygame.QUIT:
-    pass
-pygame.quit()
-
-# Удаляем за собой файл с изображением.
-os.remove(map_file)
+ch = True
+while True:
+    for event in pygame.event.get():
+        if event.type == pygame.QUIT:
+            os.remove(map_file)
+            sys.exit(0)
+        if event.type == pygame.KEYDOWN:
+            if event.key == pygame.K_PAGEDOWN:
+                z -= 1
+                if z < 0:
+                    z = 1
+                ch = True
+            if event.key == pygame.K_PAGEUP:
+                z += 1
+                if z > 17:
+                    z = 17
+                ch = True
+    if ch:
+        draw()
+    pygame.display.flip()
